dfu_progression_project/requirements.txt
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.12.0
streamlit>=1.30.0
pillow>=9.5.0
omegaconf>=2.3.0
hydra-core>=1.3.0
torchmetrics>=1.0.0
pandas>=2.0.0
